logger/requirements.txt
requests==2.21.0
lxml==4.3.1
Pillow==5.2.0